+++ paymentsheet/src/test/java/com/stripe/android/customersheet/CustomerSheetScreenshotTest.kt
@@ -15,6 +15,7 @@ import com.stripe.android.model.CardBrand
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodFixtures
 import com.stripe.android.payments.bankaccount.CollectBankAccountLauncher
+import com.stripe.android.payments.financialconnections.FinancialConnectionsAvailability
 import com.stripe.android.paymentsheet.forms.FormFieldValues
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.paymentdatacollection.FormArguments
@@ -76,6 +77,7 @@ internal class CustomerSheetScreenshotTest {
         onUpdatePrimaryButtonUIState = { },
         onError = { },
         setAsDefaultPaymentMethodEnabled = false,
+        financialConnectionsAvailability = FinancialConnectionsAvailability.Full,
         setAsDefaultMatchesSaveForFutureUse = false,
     )